+++ src/main/java/com/spacestar/back/swipe/domain/Swipe.java
@@ -2,7 +2,7 @@ package com.spacestar.back.swipe.domain;
 
 import com.spacestar.back.global.GlobalTime;
 import com.spacestar.back.swipe.SwipeStatus;
-import com.spacestar.back.swipe.dto.SwipeReqDto;
+import com.spacestar.back.swipe.dto.req.SwipeReqDto;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
