+++ src/main/java/com/spacestar/back/swipe/dto/res/SwipeResDto.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.swipe.dto;
+package com.spacestar.back.swipe.dto.res;
 
 import lombok.Getter;
