+++ src/main/java/com/spacestar/back/swipe/service/SwipeService.java
@@ -1,7 +1,6 @@
 package com.spacestar.back.swipe.service;
 
-import com.spacestar.back.swipe.dto.SwipeReqDto;
-import org.springframework.data.domain.Pageable;
+import com.spacestar.back.swipe.dto.req.SwipeReqDto;
 
 public interface SwipeService {
