+++ src/main/java/com/spacestar/back/swipe/vo/res/SwipeResVo.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.swipe.vo;
+package com.spacestar.back.swipe.vo.res;
 
 import lombok.Getter;
